Analysing return of a specific stock over a definite period of time
There are two modules for fetching the stock ticker data and for charting the returns over a specific number of days customized by the user. By running the module analysis.py enter the ticker to analyse and enter the time frame where the returns should be calculated. Ticker and the no of period is asked by the program itself.